EnglishText.jl Documentation

EnglishText.jl Documentation

Indefinite Article Selection

Word Representation of Numbers

Quantities and Pluralization

Lists of Nouns and Adjectives

Parsing Sentences

Indefinite Article Selection

EnglishText.Articulate.indefinite — Function.

indefinite(word)

Determine the correct indefinite article, from “a” or “an”, for the given noun.

julia> using EnglishText

julia> indefinite("hour")
"an"

julia> indefinite("hand")
"a"

source

Word Representation of Numbers

EnglishText.Numeric.english — Function.

Convert $n$ to English, given that $0 \le n < 10^{66}$.

julia> using EnglishText

julia> english(16)
"sixteen"

source

EnglishText.Numeric.unenglish — Function.

unenglish(T <: Integer, data::AbstractString) → T

Convert data to an integral type. This function has the guarantee that unenglish(Int, english(x)) == x, modulo any type differences. It is not guaranteed to work well or throw exceptions on other inputs.

julia> using EnglishText

julia> unenglish(Int, "sixteen")
16

source

Quantities and Pluralization

EnglishText.Quantities.ItemQuantity — Type.

ItemQuantity(n::Integer, item::AbstractString)

Represents a quantity of n occurrences of item.

julia> using EnglishText

julia> ItemQuantity(2, "apple")
2 apples

julia> ItemQuantity(1, "standard canine")
1 standard canine

source

EnglishText.Pluralize.pluralize — Function.

pluralize(word; classical=true)

Pluralize a word (given in canonical capitalization) using heuristics and lists of exceptions.

julia> using EnglishText

julia> pluralize("fox")
"foxes"

source

Lists of Nouns and Adjectives

EnglishText.ItemLists.ItemList — Constant.

ItemList(objects, connective=Sum())

A list of items or adjectives, which supports printing in standard English format.

julia> using EnglishText

julia> ItemList(["apples", "oranges"])
apples and oranges

julia> ItemList(["animal", "plant"], Disjunction())
animal or plant

julia> ItemList(["red", "blue", "white"], Conjunction())
red, blue, and white

source

Parsing Sentences

EnglishText.Text.sentences — Function.

sentences(text::AbstractString)

Return an iterable over the Sentences contained within text. Sentences are identified naïvely; that is, every full stop, exclamation mark, or question mark is considered to delimit a sentence. This is of course prone to error, as some full stops are used for abbreviations and not for delimiting sentences.

julia> using EnglishText

julia> for s in sentences("Hi! Iterate over sentences. OK?")
           println(s)
       end
Hi!
Iterate over sentences.
OK?

source